Create 1.17
Sites like Zillow get input about house prices from a database and provide nice summaries for readers. Write a program with two inputs, current price and last month's price (both integers). Then, output a summary listing the price, the change since last month, and the estimated monthly mortgage computed as (currentPrice * 0.051) / 12. End the last output with a newline.

Ex: If the input is:

200000 210000
the output is:

This house is $200000. The change is $-10000 since last month.
The estimated monthly mortgage is $850.0.